translation unit for sumfunc class
