+++ namestnikov.kirill/S5/main.cpp
@@ -4,42 +4,7 @@
 #include <utility>
 #include <map>
 #include "tree.hpp"
-
-namespace namestnikov
-{
-  class SumFunc
-  {
-  public:
-    SumFunc();
-    void operator()(const std::pair< const int, std::string > & data);
-    int getResult() const;
-    std::string getLine() const;
-  private:
-    int result_;
-    std::string line_;
-  };
-}
-
-namestnikov::SumFunc::SumFunc():
-  result_(0),
-  line_("")
-{}
-
-int namestnikov::SumFunc::getResult() const
-{
-  return result_;
-}
-
-std::string namestnikov::SumFunc::getLine() const
-{
-  return line_;
-}
-
-void namestnikov::SumFunc::operator()(const std::pair< const int, std::string > & data)
-{
-  result_ += data.first;
-  line_ = line_ + data.second + " ";
-}
+#include "sum_func.hpp"
 
 void ascending(int & result, std::string & line, namestnikov::Tree< int, std::string > & map)
 {
