+++ namestnikov.kirill/S5/sum_func.cpp
@@ -1,0 +1,22 @@
+#include "sum_func.hpp"
+
+namestnikov::SumFunc::SumFunc():
+  result_(0),
+  line_("")
+{}
+
+int namestnikov::SumFunc::getResult() const
+{
+  return result_;
+}
+
+std::string namestnikov::SumFunc::getLine() const
+{
+  return line_;
+}
+
+void namestnikov::SumFunc::operator()(const std::pair< const int, std::string > & data)
+{
+  result_ += data.first;
+  line_ = line_ + data.second + " ";
+}
